add CL fontsize argument4

+++ jupyterthemes/__init__.py
@@ -20,7 +20,7 @@ THEMES_PATH = HOME + '/.jupyter-themes'
 DEFAULT_PROFILE = 'default'
 DEFAULT_FONTSIZE='13'
 DEFAULT_TOOLBAR_STRING='div#maintoolbar {display: none !important;}'
-DEFAULT_FONTSIZE_STRING=".cm-s-ipython.CodeMirror {font-size: %spt !important;}" % DEFAULT_FONTSIZE
+DEFAULT_FONTSIZE_STRING=".cm-s-ipython.CodeMirror {font-size:%spt !important;}" % DEFAULT_FONTSIZE
 
 def get_themes():
     """ return list of available themes """
@@ -63,12 +63,12 @@ def install_path(profile=None):
     return paths
 
 
-def install_theme(name, profile=None, toolbar=False, fontsize='13'):
+def install_theme(name, profile=None, update_properties=False, toolbar=False, fontsize='13'):
     """ copy given theme to theme.css and import css in custom.css """
 
     source_path = glob('%s/%s.css' % (THEMES_PATH, name))[0]
     paths = install_path(profile)
-    FONTSIZE_STRING=".cm-s-ipython.CodeMirror {font-size: %spt !important;}" % fontsize
+    FONTSIZE_STRING=".cm-s-ipython.CodeMirror {font-size:%spt !important;}" % fontsize
     for i, target_path in enumerate(paths):
         # -- install theme
         themecss_path = '%s/theme.css' % target_path
@@ -82,19 +82,20 @@ def install_theme(name, profile=None, toolbar=False, fontsize='13'):
             if not 'theme.css' in ' '.join(customcss.readlines()):
                 customcss.seek(0, os.SEEK_END)
                 customcss.write("\n@import url('theme.css');")
-        fh, abs_path = mkstemp()
-        with open(abs_path, 'w') as cssfile:
-            with open(customcss_path) as old_file:
-                for line in old_file:
-                    if toolbar:
-                        # -- enable toolbar if requested
-                        RESTORE_TOOLBAR='/*'+DEFAULT_TOOLBAR_STRING+'*/'
-                        cssfile.write(line.replace(DEFAULT_TOOLBAR_STRING,RESTORE_TOOLBAR))
-                    # -- set CodeCell fontsize
-                    cssfile.write(line.replace(DEFAULT_FONTSIZE_STRING, FONTSIZE_STRING))
-        os.close(fh)
-        os.remove(customcss_path)
-        shutil.move(abs_path, customcss_path)
+        if update_properties:
+            fh, abs_path = mkstemp()
+            with open(abs_path, 'w') as cssfile:
+                with open(customcss_path) as old_file:
+                    for line in old_file:
+                        if toolbar:
+                            # -- enable toolbar if requested
+                            RESTORE_TOOLBAR='/*'+DEFAULT_TOOLBAR_STRING+'*/'
+                            cssfile.write(line.replace(DEFAULT_TOOLBAR_STRING,RESTORE_TOOLBAR))
+                        # -- set CodeCell fontsize
+                        cssfile.write(line.replace(DEFAULT_FONTSIZE_STRING, FONTSIZE_STRING))
+            os.close(fh)
+            os.remove(customcss_path)
+            shutil.move(abs_path, customcss_path)
 
 
 def reset_default(profile=None):
@@ -130,6 +131,7 @@ def main():
                         help="set the profile, defaults to %s" % DEFAULT_PROFILE)
     args = parser.parse_args()
 
+    update=False
     if args.list:
         themes = get_themes()
         print("Themes in %s" % THEMES_PATH)
@@ -140,7 +142,15 @@ def main():
         if args.theme not in themes:
             print("Theme %s not found. Available: %s"%(args.theme, ' '.join(themes)))
             exit(1)
-        install_theme(args.theme, profile=args.profile, toolbar=args.toolbar, fontsize=str(args.fontsize))
+        if args.toolbar:
+            update=True
+            print("Enabling toolbar")
+        else:
+            print("Toolbar is disabled. Set -T to enable")
+        if args.fontsize!=DEFAULT_FONTSIZE:
+            update=True
+
+        install_theme(args.theme, profile=args.profile, toolbar=args.toolbar, fontsize=str(args.fontsize), update_properties=update)
         exit(0)
     if args.toolbar:
         print("Enabling toolbar")
